refactor: create layout and AiBar components, implement in _app with toast error handling
