+++ frontend/package.json
@@ -14,7 +14,8 @@
         "react": "18.2.0",
         "react-dom": "18.2.0",
         "react-icons": "^5.5.0",
-        "react-paginate": "^8.3.0"
+        "react-paginate": "^8.3.0",
+        "react-toastify": "^11.0.5"
     },
     "devDependencies": {
         "autoprefixer": "^10.4.21",
